+++ lib/strong_presenter/permissions.rb
@@ -2,56 +2,61 @@ module StrongPresenter
   # @private
   #
   #   Storage format:
-  #     The permissions object is shared by collection presenters with its constituent presenters,
-  #     and it is also shared with all of its associations. Each attribute path is stored as an array
-  #     of symbols in a Set. There is one top level presenter - the one which initialized the
-  #     Permissions object.
+  #     Each attribute path is stored as an array of objects, usually strings in a Set. For indifferent access, 
+  #     all symbols are converted to strings.
   #   
-  #     When a presenter checks for permissions, the attribute path relative to the top
-  #     presenter is prepended to each attribute path, and its existence checked in the Set.
-  #
   #     Arguments can also be part of permissions control. They are simply additional elements in the attribute path array,
-  #     and need not be symbols. If they are symbols, there is no way for Permissions to know whether they
-  #     are part of the attribute path, or additional arguments. Only the presenter knows that.
+  #     and need not be strings or symbols. Permitting a string or symbol argument automatically permits both.
+  #
+  #     When checking if paths with tainted strings/elements are permitted, only exact matches are allowed
   class Permissions
 
-    # Checks whether everything is permitted.
-    #
-    # @return [Boolean]
-    def complete?
-      permitted_paths.include? [] and ((@permitted_paths = Set[[]] if permitted_paths.count > 1) or true)
+    def prefix_path
+      @prefix_path || []
     end
 
-    # Permits everything
-    #
+    # Initialize, optionally with link to permitted paths, and the prefix to that (with copy on write semantics)
+    def initialize(permissions = nil, prefix_path = [])
+      unless permissions.nil?
+        @permitted_paths = permissions.permitted_paths
+        @prefix_path = permissions.prefix_path + canonicalize(prefix_path) # copy on write
+      end
+    end
+
+    # Checks whether everything is permitted. Considers :*, which permits all methods
+    # but not association methods to be complete.
+    # @return [Boolean]
+    def complete?
+      permitted? prefix_path + [:*]
+    end
+
+    # Permits wildcard method, but not association methods.
     # @return self
     def permit_all!
-      permitted_paths.clear
-      permitted_paths << []
+      copy_on_write!
+      permitted_paths << [:*]
       self
     end
 
-    # @overload permitted? prefix_path = nil, attribute_path
-    #
     # Checks if the attribute path is permitted. This is the case if
     # any array prefix has been permitted.
-    #
-    # @param [Symbol, Array<Symbol>] prefix_path
-    # @param [Symbol, Array<Symbol,Object>] attribute_path
+    # @param [Object, Array<Object>] prefix_path
+    # @param [Object, Array<Object>] attribute_path
     # @return [Boolean]
-    def permitted? prefix_path, attribute_path = nil
-      raw_permitted? Array(prefix_path), Array(attribute_path)
+    def permitted? attribute_path
+      attribute_path = canonicalize(attribute_path)
+      return true if permitted_paths.include? prefix_path + attribute_path # exact match
+      !path_tainted?(attribute_path) and permitted_by_wildcard?(prefix_path + attribute_path)  # wildcard match only if not tainted
     end
 
     # Selects the attribute paths which are permitted.
-    #
     # @param [Array] prefix_path
     #   namespace in which each of the given attribute paths are in
-    # @param [[Symbol, Array<Symbol,Object>]*] *attribute_paths
+    # @param [[Object, Array<Object>]*] *attribute_paths
     #   each attribute path is a symbol or array of symbols
-    # @return [Array<Symbol, Array<Symbol,Object>>] array of attribute paths permitted
-    def select_permitted prefix_path, *attribute_paths
-      raw_select_permitted Array(prefix_path), nested_array(attribute_paths)
+    # @return [Array<Object, Array<Object>>] array of attribute paths permitted
+    def select_permitted *attribute_paths
+      attribute_paths.select { |attribute_path| permitted?(attribute_path) }
     end
 
     # Rejects the attribute paths which are permitted. Opposite of select_permitted.
@@ -59,37 +64,37 @@ module StrongPresenter
     #
     # @param [Array] prefix_path
     #   namespace in which each of the given attribute paths are in
-    # @param [[Symbol, Array<Symbol,Object>]*] *attribute_paths
-    #   each attribute path is a symbol or array of symbols
-    # @return [Array<Symbol, Array<Symbol,Object>>] array of attribute paths remaining
-    def reject_permitted prefix_path, *attribute_paths
-      raw_reject_permitted Array(prefix_path), nested_array(attribute_paths)
+    # @param [[Object, Array<Object>]*] *attribute_paths
+    #   each attribute path is an object(string) or array
+    # @return [Array<Object, Array<Object>>] array of attribute paths remaining
+    def reject_permitted *attribute_paths
+      attribute_paths.reject { |attribute_path| permitted?(attribute_path) }
     end
 
     # Permits some attribute paths
     #
-    # @param [Array<Symbol>] prefix_path
+    # @param [Array] prefix_path
     #   path to prepend to each attribute path
-    # @param [[Symbol, Array<Symbol,Object>]*] *attribute_paths
-    def permit prefix_path, *attribute_paths
-      prefix_path = Array(prefix_path)
+    # @param [[Object, Array]*] *attribute_paths
+    def permit *attribute_paths
+      copy_on_write!
       # don't permit if already permitted
-      raw_reject_permitted(prefix_path, nested_array(attribute_paths)).each do |attribute_path|
-        permitted_paths << prefix_path + attribute_path
+      reject_permitted(*attribute_paths).each do |attribute_path|
+        permitted_paths << canonicalize(attribute_path) # prefix_path = [] because of copy on write
       end
       self
     end
 
-    # Merges the permissions from another Permissions object
-    #
-    # @param [Permissions] permissions
-    # @param [Array<Symbol>] prefix
-    #   prefix to prepend to paths in permissions
-    # @return self
-    def merge permissions, prefix = []
-      permitted_paths.merge permissions.permitted_paths.map{|path| prefix+path} if permissions.is_a? self.class
-      self
-    end
+#    # Merges the permissions from another Permissions object
+#    #
+#    # @param [Permissions] permissions
+#    # @param [Array<Symbol>] prefix
+#    #   prefix to prepend to paths in permissions
+#    # @return self
+#    def merge permissions, prefix = []
+#      permitted_paths.merge permissions.permitted_paths.map{|path| prefix+path} if permissions.is_a? self.class
+#      self
+#    end
 
     protected
 
@@ -98,40 +103,47 @@ module StrongPresenter
     end
 
     private
-    # We trust path parameters are arrays
-    def raw_permitted? prefix_path, attribute_path = nil # const - does not alter arguments
-      return true if complete?
-      permitted_partial?([], prefix_path + Array(attribute_path))
+    def copy?
+      !@prefix_path.nil?
     end
 
-    def raw_reject_permitted prefix_path, attribute_paths # const - does not alter arguments
-      return [] if raw_permitted? prefix_path
-      attribute_paths.reject do |attribute_path|
-        permitted_partial? prefix_path.dup, attribute_path
+    def copy_on_write! # makes a copy if required
+      if prefix_path == []
+        @permitted_paths = permitted_paths.dup
+      elsif !prefix_path.nil?
+        @permitted_paths, old_set = Set.new, permitted_paths
+        old_set.each do |path|
+          @permitted_paths << path[prefix_path.size...path.size] if path[0...prefix_path.size] == prefix_path
+        end
       end
+      prefix_path = nil
     end
 
-    def raw_select_permitted prefix_path, attribute_paths # const - does not alter arguments
-      return attribute_paths if raw_permitted? prefix_path
-      attribute_paths.select do |attribute_path|
-        permitted_partial? prefix_path.dup, attribute_path
+    def path_tainted? attribute_path
+      attribute_path.tainted? or attribute_path.any? { |element| element.tainted? }
+    end
+
+    # Caution: Will mutate path
+    def permitted_by_wildcard? path
+      unless path.empty?
+        path[-1] = :*
+        return true if permitted_paths.include? path
       end
-    end
-
-    # For internal use, checks if permitted explicitly by a subpath of [prefix_path, attribute_partial+]
-    # where attribute_partial is a prefix of at least one symbol from attribute_part
-    # Caution: Will mutate prefix_path
-    def permitted_partial? prefix_path, attribute_path
-      !!Array(attribute_path).detect do |attr|
-        break unless attr.is_a? Symbol
-        prefix_path << attr
-        permitted_paths.include? prefix_path
+      until path.empty?
+        path[-1] = :**
+        return true if permitted_paths.include? path
+        path.pop
       end
+      false
     end
 
-    # Ensures that every array element is an array
-    def nested_array array
-      array.map{|e|Array(e)}
+    # Converts symbols to strings (except for wildcard symbol)
+    def canonicalize array
+      array = Array(array)
+      canonical_array = array.map{|e|e.is_a?(Symbol) ? e.to_s : e}
+      canonical_array[-1] = array.last if [:*, :**].include? array.last
+      canonical_array.taint if array.tainted?
+      canonical_array
     end
 
   end